# docs/swagger.yaml
basePath: /
definitions:
  models.Admin:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.FreelanceJob:
    properties:
      budget:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  models.Freelancer:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      rating:
        type: integer
      skills:
        type: string
      status:
        type: string
    type: object
  models.Order:
    properties:
      address_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      payment_method:
        type: string
      qr_code:
        type: string
      service_id:
        type: integer
      status:
        type: string
      user_id:
        type: integer
    type: object
  models.OrderDetail:
    properties:
      amount:
        type: integer
      description:
        type: string
      id:
        type: integer
      order_id:
        type: integer
    type: object
  models.PaymentCard:
    properties:
      card_holder:
        type: string
      card_number:
        type: string
      cvv:
        type: string
      expiry_date:
        type: string
    type: object
  models.PaymentRequest:
    properties:
      address_id:
        type: integer
      payment_card:
        $ref: '#/definitions/models.PaymentCard'
      payment_method:
        type: string
      service_id:
        type: integer
    required:
    - address_id
    - payment_method
    - service_id
    type: object
  models.PaymentResponse:
    properties:
      qr_code:
        type: string
      status:
        type: string
    type: object
  models.ProductService:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      status:
        type: string
    type: object
  models.Service:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  models.ShopLocation:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for managing services, admins, freelancers, and
    products
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PALAP Backend API
  version: "1.0"
paths:
  /api/v1/get_freelance_job:
    get:
      description: Get list of available freelance jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FreelanceJob'
            type: array
      summary: Get Freelance Jobs
  /api/v1/get_order_details:
    get:
      description: Get details for a specific order
      parameters:
      - description: Order ID
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrderDetail'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Order Details
  /api/v1/get_orders:
    get:
      description: Get all orders for authenticated user
      parameters:
      - description: User Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Orders
  /api/v1/manage_admin:
    get:
      description: Get list of all admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Admin'
            type: array
      summary: Get All Admins
    post:
      consumes:
      - application/json
      description: Create new admin
      parameters:
      - description: Admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create Admin
  /api/v1/manage_admin/:id:
    delete:
      description: Delete admin by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Admin
    get:
      description: Get admin details by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Admin by ID
    put:
      consumes:
      - application/json
      description: Update existing admin
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Admin
  /api/v1/manage_freelance:
    get:
      description: Get list of all freelancers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Freelancer'
            type: array
      summary: Get Freelancers
    post:
      consumes:
      - application/json
      description: Create new freelancer (Admin only)
      parameters:
      - description: Freelancer details
        in: body
        name: freelancer
        required: true
        schema:
          $ref: '#/definitions/models.Freelancer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Freelancer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create Freelancer
  /api/v1/manage_freelance/:id:
    delete:
      description: Delete freelancer by ID
      parameters:
      - description: Freelancer ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Freelancer
    get:
      description: Get freelancer details by ID
      parameters:
      - description: Freelancer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Freelancer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Freelancer by ID
    put:
      consumes:
      - application/json
      description: Update existing freelancer
      parameters:
      - description: Freelancer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated freelancer details
        in: body
        name: freelancer
        required: true
        schema:
          $ref: '#/definitions/models.Freelancer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Freelancer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Freelancer
  /api/v1/manage_product_and_service:
    get:
      description: Get list of all products and services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductService'
            type: array
      summary: Get Products/Services
    post:
      consumes:
      - application/json
      description: Create new product or service
      parameters:
      - description: Product/Service details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductService'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductService'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create Product/Service
  /api/v1/manage_product_and_service/:id:
    delete:
      description: Delete product/service by ID
      parameters:
      - description: Product/Service ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Product/Service
    get:
      description: Get product/service details by ID
      parameters:
      - description: Product/Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductService'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Product/Service by ID
    put:
      consumes:
      - application/json
      description: Update existing product/service
      parameters:
      - description: Product/Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product/service details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductService'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Product/Service
  /api/v1/manage_shop_location:
    get:
      description: Get list of all shop locations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ShopLocation'
            type: array
      summary: Get Shop Locations
    post:
      consumes:
      - application/json
      description: Create new shop location
      parameters:
      - description: Shop location details
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/models.ShopLocation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ShopLocation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create Shop Location
  /api/v1/manage_shop_location/:id:
    delete:
      description: Delete shop location by ID
      parameters:
      - description: Shop Location ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Shop Location
    get:
      description: Get shop location details by ID
      parameters:
      - description: Shop Location ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShopLocation'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Shop Location by ID
    put:
      consumes:
      - application/json
      description: Update existing shop location
      parameters:
      - description: Shop Location ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated shop location details
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/models.ShopLocation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShopLocation'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Shop Location
  /api/v1/payment:
    post:
      consumes:
      - application/json
      description: Process payment with card or QR code
      parameters:
      - description: User Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/models.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Payment
  /api/v1/registration:
    post:
      consumes:
      - application/json
      description: Register new freelancer
      parameters:
      - description: Freelancer details
        in: body
        name: freelancer
        required: true
        schema:
          $ref: '#/definitions/models.Freelancer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Freelancer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Freelancer Registration
  /api/v1/service:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get all service
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"